# server_source/reports/openapi_manual_export.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Murim Inn Rebuild Server",
    "version": "0.4.0"
  },
  "paths": {
...
    "/hunters": {
      "get": {
        "tags": [
          "Hunter"
        ],
        "summary": "List Hunters",
        "operationId": "list_hunters_hunters_get",
        "parameters": [
          {
            "name": "accountId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Accountid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Hunter"
        ],
        "summary": "Upsert Hunter",
        "operationId": "upsert_hunter_hunters_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Hunter"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/hunters/{hunter_id}": {
      "get": {
        "tags": [
          "Hunter"
        ],
        "summary": "Get Hunter",
        "operationId": "get_hunter_hunters__hunter_id__get",
        "parameters": [
          {
            "name": "hunter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Hunter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Hunter"
        ],
        "summary": "Delete Hunter",
        "operationId": "delete_hunter_hunters__hunter_id__delete",
        "parameters": [
          {
            "name": "hunter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Hunter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/hunters/{hunter_id}/recruit": {
      "post": {
        "tags": [
          "Hunter"
        ],
        "summary": "Recruit Hunter",
        "description": "Assign random MBTI on recruit (server-side) if hunter exists and MBTI is NONE/empty.",
        "operationId": "recruit_hunter_hunters__hunter_id__recruit_post",
        "parameters": [
          {
            "name": "hunter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Hunter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/hunters/{hunter_id}/promote": {
      "post": {
        "tags": [
          "Hunter"
        ],
        "summary": "Promote Hunter",
        "description": "Apply a promotion node to a hunter.\n\nRules:\n- Node must exist.\n- Node parent must match the last node in hunter.promotionPath (or be None for root).\n- choiceGroup can only be taken once.\n- statBonus is applied cumulatively to hunter base stats (hp/atk/defense).\n- jobId is updated to node.jobId.",
        "operationId": "promote_hunter_hunters__hunter_id__promote_post",
        "parameters": [
          {
            "name": "hunter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Hunter Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HunterPromoteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HunterPromoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/hunters/{hunter_id}/equip": {
      "post": {
        "tags": [
          "Hunter"
        ],
        "summary": "Equip Items",
        "description": "Equip item ids for a hunter with validation.\n\nValidation:\n- All itemIds must exist.\n- Only one item per slot.\n- Item seasonId must match hunter.seasonId.\n- Item tierId must be <= hunter.tierId (T1 <= T2 ...).",
        "operationId": "equip_items_hunters__hunter_id__equip_post",
        "parameters": [
          {
            "name": "hunter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Hunter Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HunterEquipRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HunterEquipResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/hunters/{hunter_id}/tier-up": {
      "post": {
        "tags": [
          "Hunter"
        ],
        "summary": "Tier Up Hunter",
        "description": "영웅 티어 상승 (T1 -> T2 -> ...). tier_defs에 정의된 티어만 허용.",
        "operationId": "tier_up_hunter_hunters__hunter_id__tier_up_post",
        "parameters": [
          {
            "name": "hunter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Hunter Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HunterTierUpRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HunterTierUpResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Hunter": {
        "properties": {
          "hunterId": {
            "type": "string",
            "title": "Hunterid"
          },
          "accountId": {
            "type": "string",
            "title": "Accountid"
          },
          "slotIndex": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Slotindex",
            "default": 0
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "jobId": {
            "type": "string",
            "title": "Jobid",
            "default": "novice"
          },
          "level": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Level",
            "default": 1
          },
          "exp": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Exp",
            "default": 0
          },
          "powerScore": {
            "type": "number",
            "title": "Powerscore",
            "default": 0.0
          },
          "hp": {
            "type": "number",
            "title": "Hp",
            "default": 0.0
          },
          "atk": {
            "type": "number",
            "title": "Atk",
            "default": 0.0
          },
          "defense": {
            "type": "number",
            "title": "Defense",
            "default": 0.0
          },
          "gold": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Gold",
            "default": 0
          },
          "gems": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Gems",
            "default": 0
          },
          "mbti": {
            "type": "string",
            "title": "Mbti",
            "default": "NONE"
          },
          "tierId": {
            "type": "string",
            "title": "Tierid",
            "default": "T1"
          },
          "seasonId": {
            "type": "string",
            "title": "Seasonid",
            "default": "S1"
          },
          "promotionPath": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Promotionpath"
          },
          "equippedItemIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Equippeditemids"
          },
          "skillsUnlocked": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skillsunlocked"
          }
        },
        "type": "object",
        "required": [
          "hunterId",
          "accountId"
        ],
        "title": "Hunter"
      },
      "HunterEquipRequest": {
        "properties": {
          "equippedItemIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Equippeditemids",
            "description": "List of equipped item ids"
          }
        },
        "type": "object",
        "title": "HunterEquipRequest"
      },
      "HunterEquipResponse": {
        "properties": {
          "hunter": {
            "$ref": "#/components/schemas/Hunter"
          },
          "itemMultipliers": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Itemmultipliers"
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "hunter"
        ],
        "title": "HunterEquipResponse"
      },
      "HunterPromoteRequest": {
        "properties": {
          "nodeId": {
            "type": "string",
            "title": "Nodeid",
            "description": "Promotion node id to apply"
          }
        },
        "type": "object",
        "required": [
          "nodeId"
        ],
        "title": "HunterPromoteRequest"
      },
      "HunterPromoteResponse": {
        "properties": {
          "hunter": {
            "$ref": "#/components/schemas/Hunter"
          },
          "appliedStatBonus": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Appliedstatbonus"
          },
          "unlockedSkills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unlockedskills"
          },
          "promotionMultiplier": {
            "type": "number",
            "title": "Promotionmultiplier",
            "default": 1.0
          }
        },
        "type": "object",
        "required": [
          "hunter"
        ],
        "title": "HunterPromoteResponse"
      },
      "HunterTierUpRequest": {
        "properties": {
          "tierId": {
            "type": "string",
            "title": "Tierid"
          }
        },
        "type": "object",
        "required": [
          "tierId"
        ],
        "title": "HunterTierUpRequest"
      },
...
    }
  }
}